- ingame HUD (supports a health bar and indication when a player is eliminated)
- minion shield now forms the shield consecutively in a circle now (doesnt take into account the manual change of the meat shield rotation based off the user input)

+++ Sector XII/Assets/Magicide/Scripts/Characters/Character.cs
@@ -86,10 +86,10 @@ public class Character : MonoBehaviour {
         return _Health;
     }
 
-    public virtual void Damage(int amount) {
+    public virtual void Damage(float amount) {
 
         // Damage character based on amount passed through
-        _Health -= amount;
+        _Health -= (int)amount;
 
         // Returns TRUE if character has no health
         if (_Health <= 0) {
